Move 'InvalidMetadataValueException' and 'NoSuchMetadataFieldException' to 'exceptions' directory.
