+++ Kitodo/src/main/java/org/kitodo/exceptions/NoSuchMetadataFieldException.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package org.kitodo.production.forms.dataeditor;
+package org.kitodo.exceptions;
 
 import java.util.Arrays;
