+++ Kitodo/src/main/java/org/kitodo/production/forms/dataeditor/MetadataTableRow.java
@@ -22,6 +22,8 @@ import org.kitodo.api.MdSec;
 import org.kitodo.api.Metadata;
 import org.kitodo.api.dataeditor.rulesetmanagement.Domain;
 import org.kitodo.api.dataformat.IncludedStructuralElement;
+import org.kitodo.exceptions.InvalidMetadataValueException;
+import org.kitodo.exceptions.NoSuchMetadataFieldException;
 import org.kitodo.production.helper.Helper;
 
 abstract class MetadataTableRow implements Serializable {
