+++ Kitodo/src/main/java/org/kitodo/production/forms/dataeditor/TextMetadataTableRow.java
@@ -33,6 +33,8 @@ import org.kitodo.api.Metadata;
 import org.kitodo.api.MetadataEntry;
 import org.kitodo.api.dataeditor.rulesetmanagement.Domain;
 import org.kitodo.api.dataeditor.rulesetmanagement.SimpleMetadataViewInterface;
+import org.kitodo.exceptions.InvalidMetadataValueException;
+import org.kitodo.exceptions.NoSuchMetadataFieldException;
 import org.kitodo.production.helper.Helper;
 
 public class TextMetadataTableRow extends SimpleMetadataTableRow implements Serializable {
